[atip] Move new added steps from suite to atip
- Before the new added steps are saved together with the tests, just move them to atip now.
